# src/asltutor/static/swagger.yaml
openapi: 3.0.0
info:
  title: ASL Tutor API
  description: Full API documentation for ASL tutor backend
  version: 1.0.0-oas3
servers:
  - url: 'http://localhost:1337/v1'
  - url: 'https://theasltutor.com/v1'
paths:
  /user/create:
    post:
      tags:
        - user
      summary: Registers a user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '200':
          description: 'Success: user registered'
        '409':
          description: 'Failed: field is already taken'
  '/user/{username}/submissions':
    get:
      tags:
        - submission
      summary: 'View a submission given a specific submission Id'
      description: 'Returns a single submission specified by a submission Id to a user. Or Returns an array of all submissions. It can be filtered based on quizId and/or moduleId and/or userId.'
      parameters:
      - name: username
        in: path
        description: 'The username of the user that the admin wants all the submissions for'
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: baba_yaga
      - name: quiz
        in: query
        description: The quiz Id that a user wants all submissions for
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: ObjectId
      - name: module
        in: query
        description: 'The username of the user that the admin wants all the submissions for'
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: ObjectId
      - name: submission
        in: query
        description: 'The Id of the submission that a user is requesting. CANNOT BE COMBINED WITH OTHER QUERIES.'
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: ObjectId
      responses:
        200:
          description: 'Success: Submission retrieved'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSubmission'
        404:
          description: 'Failed: submission not found or user does not exist'
  '/user/{username}':
    get:
      tags:
        - user
      summary: 'Get info from a user'
      description: 'Get all profile information from a user'
      parameters:
      - name: username
        in: path
        description: 'The username of the of the user who owns that profile'
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: baba_yaga
      responses:
        200:
          description: 'Success: User info retrieved'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoUser'
    post:
      tags:
        - user
      summary: 'Delete user'
      description: 'Deletes a user and all its submississions from the DB'
      parameters:
      - name: username
        in: path
        description: 'The username of the user being deleted'
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: baba_yaga
      responses:
        200:
          description: 'Success: user has been deleted'
        404:
          description: 'Failed: user not found'
...
  /module/quiz:
    post:
      tags:
        - submission
      summary: Submit a quiz
      description: Post a users quiz submission to the backend.
      requestBody:
        description: The submission object that the user is submiting
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostSubmission'
        required: true
      responses:
        '200':
          description: 'Success: Submission has been posted'
        '400':
          description: 'Failed: invalid Id(s)'
        '415':
          description: 'Failed: Content-type must be application/json'
components:
  schemas:
    InfoUser:
      description: Response object containing user info
      properties:
        username:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        DOB:
          type: string
          format: date
        email:
          type: string
          format: email
        completed_modules:
          type: array
          items:
            type: object
            properties:
              module_id:
                type: string
                format: ObjectId
              module_name:
                type: string
            example:
              module_name: module_name
              module_id: module_id
    CreateUser:
      description: User object class.
      type: object
      properties:
        username:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        DOB:
          type: string
          format: date
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        completed_modules:
          type: array
          items:
            type: object
            properties:
              module_id:
                type: string
                format: ObjectId
              module_name:
                type: string
            example:
              module_name: module_name
              module_id: module_id
      required:
        - dob
        - email
        - password
        - username
      example:
        firstname: Jon
        lastname: Wick
        password: willkillugood
        DOB: '1964-09-02'
        email: babayaga64@gmail.com
        username: baba_yaga
        completed_modules:
          - module_name: module_name
            module_id: module_id
          - module_name: module_name
            module_id: module_id
...
    GetSubmission:
      description: Content received when a user gets a submission
      type: object
      properties:
        id:
          type: string
          format: ObjectId
        user_id:
          type: string
          format: ObjectId
        quiz_id:
          type: string
          format: ObjectId
        module_id:
          type: string
          format: ObjectId
        date_completed:
          type: string
          format: date
        user_answers:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionUserAnswers'
        grade:
          type: integer
          format: int32
      example:
        quiz_id: quizId
        user_answers:
          - question_id: questionId
            user_answer: foo
          - question_id: questionId
            user_answer: bar
        date_completed: '2019-09-13 15:47:18.171396'
        grade: 6
        module_id: moduleId
        user_id: userId
        id: someId
    PostSubmission:
      description: Content received when a user gets a submission
      type: object
      properties:
        user_id:
          type: string
          format: ObjectId
        quiz_id:
          type: string
          format: ObjectId
        module_id:
          type: string
          format: ObjectId
        user_answers:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionUserAnswers'
      required:
        - user_id
        - quiz_id
        - module_id
        - user_answers
      example:
        module_id: moduleId
        user_id: userId
        quiz_id: quizId
        user_answers:
          - question_id: questionId
            user_answer: foo
          - question_id: questionId
            user_answer: bar
    SubmissionUserAnswers:
      description: Object containing user answer to a question with the questions answer Id
      type: object
      properties:
        question_id:
          type: string
          format: ObjectId
        user_answer:
          type: string
      required:
        - questionId
        - user_answer
...